+++ IrrationalNumbers/IrrationalNumbers.Logic/IBasicFunctionExpansion.cs
@@ -3,6 +3,6 @@ namespace IrrationalNumbers.Logic
 {
     public interface IBasicFunctionExpansion
     {
-        BigDecimal ExpandFunction(int wantedRemainder, double x);
+        BigDecimal ExpandFunction(int wantedRemainder, BigDecimal x);
     }
 }
